Set up sample
