+++ map_view/display-device-location/README.md
@@ -1,29 +1,40 @@
-# Display map
+# Display device location
 
-Display a map with an imagery basemap.
+Display your current position on the map, as well as switch between different types of auto pan modes.
 
-![Image of display map](DisplayMap.png)
+![Image of display device location](display-device-location.png)
 
 ## Use case
 
-The map is the fundamental building block of any GIS application and is used to specify how geographic data is organized and communicated to your users.
+When using a map within a GIS, it may be helpful for a user to know their own location within a map, whether that's to aid the user's navigation or to provide an easy means of identifying/collecting geospatial information at their location.
 
 ## How to use the sample
 
-Run the sample to view the map. Pan and zoom to navigate the map.
+Tap the button in the lower right (which starts in Stop mode). A menu will appear with the following options:
+
+* Stop - Stops the location display.
+* On - Starts the location display with no `AutoPanMode` mode set.
+* Re-Center - Starts the location display with `AutoPanMode` set to `RECENTER`.
+* Navigation - Starts the location display with `AutoPanMode` set to `NAVIGATION`.
+* Compass - Starts the location display with `AutoPanMode` set to `COMPASS_NAVIGATION`.
 
 ## How it works
 
-1. Create an ArcGIS map using a default `Basemap` such use `Basemap.createImagery()`.
-2. Create a `MapView` object to display the map.
-3. Set the map to the map view.
+1. Create a `MapView`.
+2. Get the `LocationDisplay` object by calling `getLocationDisplay()` on the map view.
+3. Use `start()` and `stop()` on the `LocationDisplay` object as necessary.
 
 ## Relevant API
 
 * ArcGISMap
-* Basemap
+* LocationDisplay
+* LocationDisplay.AutoPanMode
 * MapView
+
+## Additional information
+
+Location permissions are required for this sample.
 
 ## Tags
 
-basemap, map
+compass, GPS, location, map, mobile, navigation
