--- map_view/display-device-location/src/main/java/com/esri/samples/display_map/DisplayMapLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.display_map;
-
-/**
-* Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
-*/
-public class DisplayMapLauncher {
-
-  public static void main(String[] args) {
-    DisplayMapSample.main(args); 
-  }
-}
